Cambios:
Se añade el atributo 'role' al esquema de usuarios
Se añade en los endpoints de usuarios validaciones para que solo un usuario correctamente autenticado pueda realizar determinadas validaciones
Pequeño cambio en la estructura de datos del token
Se comenta el método 'createUser'

+++ src/api/controllers/users.controller.js
@@ -12,12 +12,19 @@ const USERmodel = require("../models/users.model");
 module.exports = {
   getAll,
   getUser,
-  createUser,
   removeUser,
   modifyUser,
 };
 
+/**
+ * @returns La lista de usuarios registrados en el sistema
+ */
 function getAll(req, res) {
+
+  if ( !isAdmin(req.locals.dataStored) ) {
+    return res.status(401).send("El usuario carece de credenciales válidas para la petición realizada");
+  }
+
   return USERmodel.find()
     .populate("reviews", {
       text: 1,
@@ -32,7 +39,15 @@ function getAll(req, res) {
     });
 }
 
+/**
+ * @returns Un usuario registrado en el sistema con el id pasado por parámetro
+ */
 function getUser(req, res) {
+
+  if ( !checkAuthUserValidity(req.locals.dataStored) ) {
+    return res.status(401).send("El usuario carece de credenciales válidas para la petición realizada");
+  }
+
   return USERmodel.findOne({ email: req.params.email })
     .then((results) => {
       return res.json(results);
@@ -42,17 +57,15 @@ function getUser(req, res) {
     });
 }
 
-function createUser(req, res) {
-  return USERmodel.create(req.body)
-    .then((results) => {
-      return res.status(201).json(results);
-    })
-    .catch((err) => {
-      return res.status(500).json(err);
-    });
-}
-
+/**
+ * @returns 
+ */
 function removeUser(req, res) {
+
+  if ( !checkAuthUserValidity(req.locals.dataStored) ) {
+    return res.status(401).send("El usuario carece de credenciales válidas para la petición realizada");
+  }
+
   return USERmodel.findByIdAndRemove(req.params.id)
     .then((results) => {
       return res.json(results);
@@ -63,6 +76,11 @@ function removeUser(req, res) {
 }
 
 function modifyUser(req, res) {
+
+  if ( !checkAuthUserValidity(req.locals.dataStored) ) {
+    return res.status(401).send("El usuario carece de credenciales válidas para la petición realizada");
+  }
+
   return USERmodel.findByIdAndUpdate(req.params.id, req.body, { new: true })
     .then((results) => {
       return res.json(results);
@@ -70,4 +88,26 @@ function modifyUser(req, res) {
     .catch((err) => {
       return res.status(500).json(err);
     });
+}
+
+/**
+ * La función verifica que el usuario tiene permiso para realizar la 
+ * request. Tendrá permiso siempre que sea userRole==="admin" o 
+ * un userRole==="user" que realiza una operación sobre su misma id
+ */
+function checkAuthUserValidity(dataStored) {
+  const userRole = dataStored.userRole;
+  const userId = dataStored.userId
+
+  return userRole === "user" && userId == req.params.id 
+}
+
+/**
+ * La función verifica que el usuario tiene userRole==="admin"
+ */
+ function isAdmin(dataStored) {
+  const userRole = dataStored.userRole;
+
+  console.log(userRole);
+  return userRole === "admin"
 }
